+++ src/Operations/Coalesce.php
@@ -15,7 +15,7 @@ class Coalesce extends Operation
             "%s%s%s%s",
             $this->operator,
             $this->preSeparator,
-            implode($this->separator, $this->operands),
+            implode($this->separator, $this->expressions),
             $this->postSeparator
         );
     }